+++ embassy-mspm0/src/lib.rs
@@ -430,22 +430,6 @@ pub(crate) mod sealed {
     pub trait Sealed {}
 }
 
-struct BitIter(u32);
-
-impl Iterator for BitIter {
-    type Item = u32;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.0.trailing_zeros() {
-            32 => None,
-            b => {
-                self.0 &= !(1 << b);
-                Some(b)
-            }
-        }
-    }
-}
-
 /// Reset cause values from SYSCTL.RSTCAUSE register.
 /// Based on MSPM0 L-series Technical Reference Manual Table 2-9 and
 /// MSPM0 G-series Technical Reference Manual Table 2-12.